Move schematype by type - created two folders documents and objects. - All schemas of type documents have been moved to documents - All schemas of type object have been moved to objects
