+++ sanity-portfolio/schemas/documents/technology.ts
@@ -1,0 +1,14 @@
+import {defineType} from 'sanity'
+
+export default defineType({
+  name: 'technology',
+  type: 'document',
+  title: 'Technologies',
+  fields: [
+    {
+      name: 'name',
+      title: 'Name',
+      type: 'string',
+    },
+  ],
+})
