+++ sanity-portfolio/schemas/index.ts
@@ -1,17 +1,18 @@
-import blockContent from './blockContent'
-import exeperience from './exeperience'
-
-import project from './project'
-import tags from './tags'
-import testinomials from './testinomials'
+import blockContent from './objects/blockContent'
+import exeperience from './documents/exeperience'
+import project from './documents/project'
+import tag from './objects/tag'
+import testinomial from './documents/testinomial'
+import technology from './documents/technology'
 
 export const schemaTypes = [
   // Document types
   project,
-  testinomials,
+  testinomial,
   exeperience,
+  technology,
 
   // Other types
   blockContent,
-  tags,
+  tag,
 ]
